backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib>=1.7.4
python-multipart>=0.0.6
boto3>=1.34.0
botocore>=1.34.0
aiofiles>=23.2.0
email-validator>=2.0.0
celery[redis]>=5.3.0
redis>=5.0.0

